// src/model/contact.model.ts
export class ContactResponse {
  id: string;
  first_name: string;
  last_name?: string;
  email?: string;
  phone?: string;
}

export class CreateContactRequest {
  first_name: string;
  last_name?: string;
  email?: string;
  phone?: string;
}

export class UpdateContactRequest {
  id: string;
  first_name: string;
  last_name?: string;
  email?: string;
  phone?: string;
}